Calculate percentages of variation accounted by synthetic axes

Calculate the percentage of total original variation accounted by syntethtic axes generated by different multivariate ordination methods.

Usage

exp_var(ordination)

Arguments

ordination

An ordination (i.e. a "prcomp", "bg_prcomp", "phy_prcomp", "pls_shapes" or "phy_pls_shapes" object).

Value

A table informing the percentages and cumulative percentages of original variation accounted by each synthetic axis of the multivariate ordination.

See also

Examples

#load data
data("shells")

#extract shapes, log sizes and species classification
species <- shells$data$species
shapes <- shells$shapes$coe
sizes <- log(shells$sizes)

#perform PCA, bgPCA, and PLS
pca <- prcomp(shapes)
bgpca <- bg_prcomp(shapes, groups = species)
pls <- pls_shapes(shapes, X = sizes)

#compare percentages of total variation accounted by the first axes when
#the analysis is unsupervised (PCA), or supervised by species (bgPCA) or
#by log size (PLS)
head(exp_var(pca))
#>     variance cummulative
#> PC1 88.65268    88.65268
#> PC2  4.96295    93.61562
#> PC3  2.97799    96.59362
#> PC4  1.24819    97.84181
#> PC5  0.67411    98.51592
#> PC6  0.47215    98.98807
exp_var(bgpca)
#>       variance cummulative
#> bgPC1 88.52150    88.52150
#> bgPC2  4.67040    93.19190
#> bgPC3  2.80097    95.99287
exp_var(pls)
#>       variance cummulative
#> PLS-1  7.86911     7.86911