Project shapes into morphospace

Project a set of shapes into an existing morphospace.

Usage

proj_shapes(shapes, mspace, pipe = TRUE, ...)

Arguments

shapes

Shape data.

mspace

An "mspace" object.

pipe

Logical; is the function being included in a pipe?

...

Further arguments passed to points().

Value

If a plot device with a morphospace is open, shapes feeded to shapes are projected into morphospace. If pipe = FALSEthose scores are returned invisibly. If pipe = TRUE the feeded mspace object is returned unchanged and invisibly.

Details

The purpose of this function is to mantain morphospace generation and sample representation as independent affairs, and to add flexibility to graphical representation of scatter points.

Examples

#load and extract relevant data, packages and information
library(magrittr)
data("wings")
shapes <- wings$shapes
cactus <- wings$data$sex
species <- wings$data$species
template <- wings$template

#generate basic morphospace, add sampled shapes
mspace(shapes, template = template, mag = 0.7, axes = c(1,2)) %>%
  proj_shapes(shapes = shapes)
#> Error in shapes_mat(shapes): object 'data2d' not found

#change colors, symbols, etc for the scatter
mspace(shapes, template = template, mag = 0.7, axes = c(1,2)) %>%
  proj_shapes(shapes = shapes[,,species == "Db"],  col = c("green"),
              pch = c(1,16)[cactus[species == "Db"]]) %>%
  proj_shapes(shapes = shapes[,,species == "Dk"],  col = c("blue"),
              pch = c(1,16)[cactus[species == "Dk"]])
#> Error in shapes_mat(shapes): object 'data2d' not found