Ptychomya shell outlines data set

Sample of intra- and interspecific shell shapes, quantified as closed outlines using 7 harmonics and elliptic Fourier analysis, from 4 species of the extinct genus Ptychomya (Bivalvia: Crassatellidae).

Usage

shells

Format

A list containing:

$shapes:

an "OutCoe" object containing a 137 x 28 matrix with the normalized Fourier coefficients describing the outline of the shell of each specimen in left lateral view.

$sizes:

a vector of length 137 containing the centroid size of each fossil shell.

$data:

a 4-column data frame with information about the taxonomic classification, ammonoid biozone, estimated geochronologic age (in million years), and geographic provenance of each fossil shell.

Details

The shells data contain data from 137 specimens belonging to 4 species of the extinct bivalve genus Ptychomya, tracking their morphological changes through a 5 million years interval from the Lower Cretaceous of west-central Argentina (approximately 140 million years ago). The data set includes the information about the extrernal shell shape (measured using Fourier coefficients from 7 harmonics), centroid size, age (both relative, taken from ammonoid biozones, and absolute, estimated using a combination of the former, absolute dates, and stratigraphic data), geographic provenance, and taxonomic classification of each fossil specimen.

References

Milla Carmona, P. S., Lazo, D. G., & Soto, I. M. (2018). Morphological evolution of the bivalve Ptychomya through the Lower Cretaceous of Argentina. Paleobiology, 44(1), 101-117.