Steinmanella shell surfaces data set

Sample of ontogenetic longitudinal shell shapes, quantified using 90 surface semilandmarks, from 7 species of the extinct genus Steinmanella (Bivalvia: Trigoniida).

shells3D

Format

A list containing:

$shapes:

a 90 x 3 x 278 array with the slid, superimposed configurations of semilandmarks describing surfaces representing right valves (3-5 per specimen).

$sizes:

a vector of length 278 containing the centroid size of each shell surface

$data:

a 4-column data frame with information about the individual id, taxonomic classification, amonoid biozone and geographic provenance of each shell surface.

$mesh_meanspec:

a 3D surface mesh stored as a "mesh3d" object (from rgl) corresponding to the shell shape closest to the sample's mean (found using geomorph::findMeanSpec()), to be warped using TPS interpolation.

Details

The shells3D data contain data from 67 specimens belonging to 7 species of the extinct bivalve genus Steinmanella, which inhabited the Early Cretaceous of west-central Argentina during a ~7 million years interval. Shapes were measured in each individual using growth lines marking the form of the animal at different ages, making a total of 278 shell surfaces describing intra- and interspecfic ontogenetic variation. This data includes the semilandmarks describing these surfaces' shapes (digitally processed to maximize signal), their centroid sizes, the individual to which they belong, as well as their relative age (ammonoid biozones), geographic provenance, and taxonomic classification. In order to improve visualization, the surface mesh of the specimen closest to the consensus of the sample, to be warped using TPS interpoltation, has also been included.

References

Milla Carmona, P. S., Lazo, D. G., & Soto, I. M. (2021). Ontogeny in the steinmanellines (Bivalvia: Trigoniida): an intra- and interspecific appraisal using the Early Cretaceous faunas from the Neuquen Basin as a case study. Paleobiology, in press.