Tyrannus tail shape data set

Sample of intra- and interspecific tail shapes, quntified using 9 landmarks, from the 13 species of the genus Tyrannus (Aves: Tyrannidae).

Usage

tails

Format

A list containing:

$shapes:

a 9 x 2 x 281 array with the symmetric component of the superimposed configurations of landmarks marking the base of the tail, as well as the tips of the two outermost rectrices of each specimen.

$sizes:

a vector of length 281 containing the centroid size of each specimen's tail.

$data:

a 3-column data frame with information about the taxonomic classification, sex and type (deep forked, DF or non-deep forked, NDF.

$tree:

a "phy" object containing the phylogenetic tree of the 13 Tyrannus species.

$links:

a list indicating how to connect the landmarks to create a wireframe.

Details

This data set contains a sample of tail shapes from the 13 species of the genus Tyrannus, two of which (T. savana and T. forficatus) display exaggeratedly elongated tails, as well as a considerable allometric variation and sexual dimorphism. The tails data set contains landmark data and centroid sizes from the tails of 281 specimens, their classification to species and sex, and the phylogenetic relationships between Tyrannus species (see Fasanelli et al. 2022 and references therein). To further help visualization of shapes, the links between landmarks have also been included.

References

Fasanelli M. N., Milla Carmona P. S., Soto I. M., & Tuero, D.T . (2022). Allometry, sexual selection and evolutionary lines of least resistance shaped the evolution of exaggerated sexual traits within the genus Tyrannus. Journal of Evolutionary Biology, 35(5), 669 - 679.