Plot morphospaces and combine them with other variables

Flexible representation of morphospaces, including their combination with other variables or a phylogeny.

plot_mspace(
  mspace,
  axes = NULL,
  links = NULL,
  template = NULL,
  x = NULL,
  y = NULL,
  nh,
  nv,
  mag,
  invax = NULL,
  adj_frame = c(1, 1),
  points = TRUE,
  models = TRUE,
  mshapes = TRUE,
  groups = TRUE,
  phylo = TRUE,
  shapeax = TRUE,
  legend = FALSE,
  cex.legend = 1,
  asp = NA,
  xlab,
  ylab,
  xlim = NULL,
  ylim = NULL,
  size.models,
  asp.models,
  rot.models,
  col.models,
  bg.models,
  lwd.models,
  alpha.models,
  cex.ldm,
  col.ldm,
  pch.points = 1,
  col.points = 1,
  bg.points = 1,
  cex.points = 1,
  density.points = TRUE,
  col.groups = 1,
  bg.groups = 1,
  pch.groups = 16,
  cex.groups = 1,
  ellipse.groups = mspace$plotinfo$ellipse.groups,
  conflev.groups = 0.95,
  lwd.groups = 1,
  lty.groups = 1,
  alpha.groups = 0,
  density.groups = TRUE,
  lwd.phylo = 1,
  lty.phylo = 1,
  col.phylo = 1,
  lwd.axis = 1,
  lty.axis = 1,
  col.axis = 1
)

Arguments

mspace

An "mspace" object created using mspace().

axes

Numeric of length 1 or 2, indicating the axes to be plotted.

links

A list with the indices of the coordinates defining the wireframe (following the format used in Morpho).

template

Either a 2-column matrix containing landmarks/semilandmarks followed by coordinates defining a curve or set of curves describing additional aspects of morphology (for 2D shape data) or a "mesh3d" object containing geometry of the structure the landmarks were placed on (for 3D shape data), corresponding to the mean shape of the sample. These will be warped using TPS interpolation to produce the set of background shell models (see build_template2d).

x

Optional vector with a non-morphometric variable to be plotted in the x axis. Alternatively, a "phylo" object can be provided.

y

Optional vector with a non-morphometric variable to be plotted in the y axis. Alternatively, a "phylo" object can be provided.

nh

Numeric; the number of shape models along the x axis.

nv

Numeric; the number of shape models along the y axis.

mag

Numeric; magnifying factor for shape models.

invax

Optional numeric indicating which of the axes provided in axes needs to be inverted (optionsare 1, 2 or c(1,2)).

adj_frame

Numeric of length 2, providing a posteriori scaling factors for the width and height of the frame, respectively.

points

Logical; whether to plot the scatter points corresponding to the sampled shapes stored in mspace$x.

models

Logical; whether to plot background shape models.

mshapes

Logical; whether to plot the scatter points corresponding to groups' mean shapes stored in mspace$gr_centroids.

groups

Logical; whether to plot the convex hulls/confidence ellipses enclosing the groups stored in mspace$gr_class.

phylo

Logical; whether to plot phylogenetic relationships stored in mspace$phylo.

shapeax

Logical; whether to plot morphometric axes stored in mspace$shape_axis.

legend

Logical; whether to show legend for groups (mspace$gr_class).

cex.legend

Numeric; size of legend labels/symbols.

xlim, ylim, xlab, ylab, asp

Standard arguments passed to the generic graphics::plot() function.

size.models

Numeric; size factor for shape models.

asp.models

Numeric; y/x aspect ratio of shape models.

rot.models

Numeric; angle (in degrees) to rotate shape models.

col.models

Color for wireframes/outlines.

bg.models

Background color for outlines/meshes.

lwd.models

Numeric; width of the lines in wireframes/outlines.

alpha.models

Numeric; transparency factor for background models (3D only).

cex.ldm

Numeric; size of landmarks/semilandmarks in the background models.

col.ldm

Color of landmarks/semilandmarks in the background models.

pch.points

Numeric; symbol of the scatter points corresponding to sampled shapes.

col.points

Color of the scatter points corresponding to sampled shapes.

bg.points

Background color of the scatter points corresponding to sampled shapes.

cex.points

Numeric; size of the scatter points corresponding to sampled shapes.

density.points

Logical; whether to add density distribution for points (univariate ordinations only).

col.groups

Color of the hulls/ellipses and/or scatter points corresponding to groups mean shapes.

bg.groups

Background color of the scatter points corresponding to groups mean shapes.

pch.groups

Numeric; symbol of the scatter points corresponding to groups mean shapes.

cex.groups

Numeric; size of the scatter points corresponding to groups mean shapes.

ellipse.groups

Logical; whether to plot confidence ellipses (if FALSE, convex hulls will be used instead).

conflev.groups

Numeric; confidence level used for confidence ellipse(s).

lwd.groups

Numeric; width of the lines in groups' ellipses/hulls.

lty.groups

Numeric; type of the lines in groups' ellipses/hulls.

alpha.groups

Numeric; transparency factor for groups' ellipses/hulls/density distributions.

density.groups

Logical; whether to add density distribution for groups (univariate ordinations only).

lwd.phylo

Numeric; width of the lines depicting phylogenetic branches.

lty.phylo

Numeric; type of the lines depicting phylogenetic branches.

col.phylo

Numeric; color of the lines depicting phylogenetic branches.

lwd.axis

Numeric; width of the lines depicting morphometric axis.

lty.axis

Numeric; type of the lines depicting morphometric axis.

col.axis

Numeric; color of the lines depicting morphometric axis.

Details

This function allows to regenerate morphospaces contained in mspace objects already in existence, either preserving the graphical attributes used during the pipeline or modifying one or more of them (so there is no need to execute the pipeline every time morphospaces need to be visualized and/or changed).

Also, plot_mspace expands the range of graphical options available beyond 'pure' morphospaces. If a numeric non-shape variable (assumed to have been measured for the same specimens in mspace$x) is feeded to one of x or y, a 'hybrid' morphospace is produced (i.e. the bivariate plot will be constructed from the combination of x or y and a morphometric axis; shape models in the background will represent variation only for the latter). If instead a "phylo" object (assumed to describe the phylogenetic relationships among tips scores stored in mspace$phylo_scores) is feeded to one of x or y, a vertical or horizontal phenogram will be deployed (the x/y axis range will correspond to branch lengths, so caution should be exercised when interpreting the output).

Examples

#load and extract relevant data, packages and information
library(magrittr)
data("tails")
shapes <- tails$shapes
species <- tails$data$species
sizes <- tails$sizes
sp_shapes <- expected_shapes(shapes, species)
sp_sizes <- cbind(tapply(sizes, species, mean))
tree <- tails$tree
links <- tails$links

#generate basic morphospace, add sampled shapes, species mean shapes, species
#classification, and phylogenetic structure
msp <- mspace(shapes, axes = c(1,2), plot = FALSE) %>%
 proj_shapes(shapes = shapes) %>%
 proj_consensus(shapes = sp_shapes) %>%
 proj_groups(groups = species) %>%
 proj_phylogeny(tree = tree)
#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet

##Plotting 'pure' morphospaces:

#plot mspace object as it is
plot_mspace(msp, axes = c(1,2),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#add colors for points, by species
plot_mspace(msp, axes = c(1,2), col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#add links
plot_mspace(msp, axes = c(1,2), links = links,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#change number and sizes of shape models in the background
plot_mspace(msp, axes = c(1,2), nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species),
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#magnify deformation and highlight landmarks
plot_mspace(msp, axes = c(1,2), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#change axes 1,2 for 1,3
plot_mspace(msp, axes = c(1,3), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#represent ranges and centroids of groups
plot_mspace(msp, axes = c(1,3), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = TRUE, mshapes = TRUE, phylo = FALSE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

#add phylogeny
plot_mspace(msp, axes = c(1,3), mag = 1.5, nh = 2, nv = 2, links = links, size.models = 0.5,
            col.points = species, col.groups = 1:nlevels(species), cex.ldm = 5, col.ldm = "red",
            points = TRUE, groups = TRUE, mshapes = TRUE, phylo = TRUE)
#> Error in eval(expr, envir, enclos): object 'msp' not found

##Plotting 'hybrid' morphospaces:

#plot size against first PC
plot_mspace(msp, x = sizes,  axes = 1, links = links, col.points = species,
            col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Centroid size")
#> Error in eval(expr, envir, enclos): object 'msp' not found

##Plotting phenograms:

#plot horizontal phenogram against PC1
plot_mspace(msp, x = tree,  axes = 1, links = links, col.points = species,
            col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, xlab = "Branch lengths")
#> Error in eval(expr, envir, enclos): object 'msp' not found

#plot horizontal phenogram against PC1
plot_mspace(msp, y = tree,  axes = 2, links = links, col.points = species,
            col.groups = 1:nlevels(species), pch.points = 16,
            points = TRUE, groups = FALSE, mshapes = FALSE, phylo = FALSE, ylab = "Branch lengths")
#> Error in eval(expr, envir, enclos): object 'msp' not found